+++ twitchAPI/__init__.py
@@ -2,3 +2,4 @@ from twitchAPI.twitch import Twitch
 from twitchAPI.webhook import TwitchWebHook
 from twitchAPI.oauth import UserAuthenticator, refresh_access_token
 import twitchAPI.types
+from twitchAPI.pubsub import PubSub