+++ twitchAPI/helper.py
@@ -14,6 +14,7 @@ from urllib.parse import urlparse, parse_qs
 
 TWITCH_API_BASE_URL = "https://api.twitch.tv/helix/"
 TWITCH_AUTH_BASE_URL = "https://id.twitch.tv/"
+TWITCH_PUB_SUB_URL = "wss://pubsub-edge.twitch.tv"
 
 
 def extract_uuid_str_from_url(url: str) -> Union[str, None]:
@@ -160,3 +161,11 @@ def fields_to_enum(data: Union[dict, list],
         return [make_dict_field_enum(d, fields, _enum, default) for d in data]
     else:
         return make_dict_field_enum(data, fields, _enum, default)
+
+
+def make_enum(data: str, _enum: Type[Enum], default: Enum) -> Enum:
+    _enum_vals = [e.value for e in _enum.__members__.values()]
+    if data in _enum_vals:
+        return _enum(data)
+    else:
+        return default