+++ twitchAPI/twitch.py
@@ -62,6 +62,7 @@ from typing import Union, List, Optional
 from .helper import build_url, TWITCH_API_BASE_URL, TWITCH_AUTH_BASE_URL, make_fields_datetime, build_scope, \
     fields_to_enum
 from datetime import datetime
+from logging import getLogger, Logger
 from .types import *
 
 
@@ -84,11 +85,14 @@ class Twitch:
     __user_auth_scope: List[AuthScope] = []
     __has_user_auth: bool = False
 
+    __logger: Logger = None
+
     auto_refresh_auth: bool = True
 
     def __init__(self, app_id: str, app_secret: str):
         self.app_id = app_id
         self.app_secret = app_secret
+        self.__logger = getLogger('twitchAPI.twitch')
 
     def __generate_header(self, auth_type: 'AuthType', required_scope: List[AuthScope]) -> dict:
         header = {"Client-ID": self.app_id}
@@ -112,9 +116,14 @@ class Twitch:
                 f'Bearer {self.__user_auth_token if self.__has_user_auth else self.__app_auth_token}'
         return header
 
+    def get_user_auth_scope(self) -> List[AuthScope]:
+        """Returns the set User auth Scope"""
+        return self.__user_auth_scope
+
     def refresh_used_token(self):
         """Refreshes the currently used token"""
         if self.__has_user_auth:
+            self.__logger.debug('refreshing user token')
             from .oauth import refresh_access_token
             self.__user_auth_token,\
                 self.__user_auth_refresh_token = refresh_access_token(self.__user_auth_refresh_token,
@@ -131,6 +140,7 @@ class Twitch:
                            retries: int = 1) -> requests.Response:
         """Make POST request with authorization"""
         headers = self.__generate_header(auth_type, required_scope)
+        self.__logger.debug(f'making POST request to {url}')
         req = None
         if data is None:
             req = requests.post(url, headers=headers)
@@ -157,6 +167,7 @@ class Twitch:
                           retries: int = 1) -> requests.Response:
         """Make PUT request with authorization"""
         headers = self.__generate_header(auth_type, required_scope)
+        self.__logger.debug(f'making PUT request to {url}')
         req = None
         if data is None:
             req = requests.put(url, headers=headers)
@@ -183,6 +194,7 @@ class Twitch:
                             retries: int = 1) -> requests.Response:
         """Make PATCH request with authorization"""
         headers = self.__generate_header(auth_type, required_scope)
+        self.__logger.debug(f'making PATCH request to {url}')
         req = None
         if data is None:
             req = requests.patch(url, headers=headers)
@@ -209,6 +221,7 @@ class Twitch:
                              retries: int = 1) -> requests.Response:
         """Make DELETE request with authorization"""
         headers = self.__generate_header(auth_type, required_scope)
+        self.__logger.debug(f'making DELETE request to {url}')
         req = None
         if data is None:
             req = requests.delete(url, headers=headers)
@@ -233,6 +246,7 @@ class Twitch:
                           retries: int = 1) -> requests.Response:
         """Make GET request with authorization"""
         headers = self.__generate_header(auth_type, required_scope)
+        self.__logger.debug(f'making GET request to {url}')
         req = requests.get(url, headers=headers)
         if self.auto_refresh_auth and retries > 0:
             if req.status_code == 401:
@@ -254,6 +268,7 @@ class Twitch:
             'grant_type': 'client_credentials',
             'scope': build_scope(self.__app_auth_scope)
         }
+        self.__logger.debug('generating fresh app token')
         url = build_url(TWITCH_AUTH_BASE_URL + 'oauth2/token', params)
         result = requests.post(url)
         if result.status_code != 200:
@@ -437,7 +452,7 @@ class Twitch:
         For detailed documentation, see here: https://dev.twitch.tv/docs/api/reference#get-bits-leaderboard
 
         :param int count: Number of results to be returned. In range 1 to 100, |default| :code:`10`
-        :param ~twitchAPI.types.TimePeriod period: Time period over which data is aggregated, default
+        :param ~twitchAPI.types.TimePeriod period: Time period over which data is aggregated, |default|
                 :const:`twitchAPI.types.TimePeriod.ALL`
         :param ~datetime.datetime started_at: Timestamp for the period over which the returned data is aggregated.
         :param str user_id: ID of the user whose results are returned; i.e., the person who paid for the Bits.
@@ -1304,7 +1319,7 @@ class Twitch:
         :param str after: Cursor for forward pagination
         :param str before: Cursor for backward pagination
         :param int first: Number of values to be returned when getting videos by user or game ID.
-                        Limit: 100. Default: 20.
+                        Limit: 100. |default| :code:`20`
         :param str language: Language of the video being queried.
         :param ~twitchAPI.types.TimePeriod period: Period during which the video was created.
         :param ~twitchAPI.types.SortMethod sort: Sort order of the videos.
@@ -1431,7 +1446,7 @@ class Twitch:
         :param str query: search query
         :param int first: Maximum number of objects to return. Maximum: 100 |default| :code:`20`
         :param str after: Cursor for forward pagination
-        :param bool live_only: Filter results for live streams only. Default: False
+        :param bool live_only: Filter results for live streams only. |default| :code:`False`
         :raises ~twitchAPI.types.UnauthorizedException: if app authentication is not set
         :raises ~twitchAPI.types.TwitchAuthorizationException: if the used authentication token became invalid
                         and a re authentication failed
@@ -1534,7 +1549,7 @@ class Twitch:
         :param str from_id: User ID of the follower
         :param str to_id: ID of the channel to be followed by the user
         :param bool allow_notifications: If true, the user gets email or push notifications (depending on the user’s
-                        notification settings) when the channel goes live. Default value is false.
+                        notification settings) when the channel goes live. |default| :code:`false`
         :raises ~twitchAPI.types.UnauthorizedException: if user authentication is not set
         :raises ~twitchAPI.types.MissingScopeException: if the user authentication is missing the required scope
         :raises ~twitchAPI.types.TwitchAuthorizationException: if the used authentication token became invalid